@@ -32,19 +32,20 @@ class _UpdateEmailWidgetState extends State<UpdateEmailWidget> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
+      backgroundColor: MyColors.greyBlue,
       appBar: PreferredSize(
         child: CommonAppbar(),
         preferredSize: Size.fromHeight(height(context) / 5),
       ),
       body: BlocConsumer<UpdateEmailBloc, UpdateEmailState>(
-        listener: (context, state) => {},
+        listener: (context, state) => {
+          if (state is ErrorUpdateEmailState)
+            {AlertMessage().alert(context, state.message)},
+        },
         builder: (context, state) {
           print("state is $state");
           if (state is UpdateEmailInitial) {
             return _buildEmailbody();
-          }
-          if (state is ErrorUpdateEmailState) {
-            return Center(child: Text(state.message));
           }
           if (state is LoadingUpdateEmailState) {
             return const Center(child: CircularProgressIndicator());
@@ -54,126 +55,131 @@ class _UpdateEmailWidgetState extends State<UpdateEmailWidget> {
             var emailToken = state.emailToken;
             return _buildCodebody(context, emailToken);
           }
-          return const Center(child: CircularProgressIndicator());
+          return _buildEmailbody();
         },
       ),
     );
   }
 
   _buildCodebody(BuildContext context, String emailToken) {
-    return BlocBuilder<UpdateEmailByCodeBloc, UpdateEmailByCodeState>(
+    return BlocConsumer<UpdateEmailByCodeBloc, UpdateEmailByCodeState>(
+      listener: (context, state) {
+        if (state is ErrorUpdateEmailByCodeState) {
+          AlertMessage().alert(context, state.message);
+        }
+        if (state is SuccessUpdateEmailByCodeState) {
+          AlertMessage().alert(context, state.message);
+          Navigator.pop(context);
+        }
+        if (state is LoadingUpdateEmailByCodeState) {
+          AlertMessage().loading(context);
+        }
+      },
       builder: (context, state) {
         print("state is $state");
         if (state is UpdateEmailByCodeInitial) {
-          return Padding(
-              padding: EdgeInsets.all(30),
-              child: Form(
-                key: _formKey,
-                child: Column(
-                  children: [
-                    Container(
-                      padding: EdgeInsets.only(left: 20),
-                      decoration: BoxDecoration(
-                        color: Colors.white,
-                        // border: Border.all(
-                        //     //color: Color.fromARGB(255, 192, 190, 190)),
-                        borderRadius: const BorderRadius.all(
-                          Radius.circular(
-                            15,
-                          ),
-                        ),
-
-                        boxShadow: [
-                          BoxShadow(
-                            color: Colors.black.withOpacity(
-                              0.1,
-                            ),
-                            spreadRadius: 1,
-                            blurRadius: 2,
-                            offset: const Offset(
-                                0, 1), // changes position of shadow
-                          ),
-                        ],
-                      ),
-                      child: TextFormField(
-                        controller: code,
-                        autofillHints: [AutofillHints.oneTimeCode],
-                        //onEditingComplete: ()=>TextInput.finishAutofillContext(),
-                        keyboardType: TextInputType.number,
-                        maxLength: 6,
-
-                        decoration: InputDecoration(
-                            counterText: '',
-                            border: InputBorder.none,
-                            labelStyle: TextStyle(
-                                color: MyColors.lightFont, fontSize: 13),
-                            labelText: 'OTP'),
-
-                        onChanged: (value) {},
-                        onSaved: (value) {
-                          newPasswordData = value;
-                        },
-                        validator: (value) {
-                          if (value!.isEmpty) {
-                            return 'Please enter New Password';
-                          }
-                          return null;
-                        },
-                      ),
-                    ),
-                    SizedBox(
-                      height: 50,
-                    ),
-                    Align(
-                        alignment: Alignment.bottomCenter,
-                        child: Column(children: [
-                          Container(
-                            width: width(context),
-                            decoration: BoxDecoration(
-                              color: MyColors.blue,
-                              borderRadius: BorderRadius.all(
-                                Radius.circular(20),
-                              ),
-                            ),
-                            //margin: EdgeInsets.all(25),
-                            child: FlatButton(
-                              child: Text(
-                                'Confirm',
-                                style: TextStyle(fontSize: 20.0),
-                              ),
-                              textColor: MyColors.white,
-                              onPressed: () {
-                                if (_formKey.currentState!.validate()) {
-                                  _formKey.currentState!.save();
-
-                                  BlocProvider.of<UpdateEmailByCodeBloc>(
-                                          context)
-                                      .add(UpdateEmailByCodeProfileEvent(
-                                          code: code.text,
-                                          emailToken: emailToken,
-                                          context: context));
-                                }
-                              },
-                            ),
-                          ),
-                        ]))
-                  ],
-                ),
-              ));
-        }
-        if (state is ErrorUpdateEmailByCodeState) {
-          return AlertMessage().alert(context, state.message);
+          return _buildEmailByCodeBody(emailToken);
         }
         if (state is LoadingUpdateEmailByCodeState) {
-          return const Center(child: CircularProgressIndicator());
+          return Center(child: CircularProgressIndicator());
         }
 
-        if (state is SuccessUpdateEmailByCodeState) {
-          return AlertMessage().alert(context, state.message);
-        }
-        return const Center(child: CircularProgressIndicator());
+        return _buildEmailByCodeBody(emailToken);
       },
     );
+  }
+
+  _buildEmailByCodeBody(String emailToken) {
+    return Padding(
+        padding: EdgeInsets.all(30),
+        child: Form(
+          key: _formKey,
+          child: Column(
+            children: [
+              Container(
+                padding: EdgeInsets.only(left: 20),
+                decoration: BoxDecoration(
+                  color: Colors.white,
+                  // border: Border.all(
+                  //     //color: Color.fromARGB(255, 192, 190, 190)),
+                  borderRadius: const BorderRadius.all(
+                    Radius.circular(
+                      15,
+                    ),
+                  ),
+
+                  boxShadow: [
+                    BoxShadow(
+                      color: Colors.black.withOpacity(
+                        0.1,
+                      ),
+                      spreadRadius: 1,
+                      blurRadius: 2,
+                      offset: const Offset(0, 1), // changes position of shadow
+                    ),
+                  ],
+                ),
+                child: TextFormField(
+                  controller: code,
+                  autofillHints: [AutofillHints.oneTimeCode],
+                  //onEditingComplete: ()=>TextInput.finishAutofillContext(),
+                  keyboardType: TextInputType.number,
+                  maxLength: 6,
+
+                  decoration: InputDecoration(
+                      counterText: '',
+                      border: InputBorder.none,
+                      labelStyle:
+                          TextStyle(color: MyColors.lightFont, fontSize: 13),
+                      labelText: 'OTP'),
+
+                  onChanged: (value) {},
+                  onSaved: (value) {
+                    newPasswordData = value;
+                  },
+                  validator: (value) {
+                    if (value!.isEmpty) {
+                      return 'Please enter New Password';
+                    }
+                    return null;
+                  },
+                ),
+              ),
+              SizedBox(
+                height: 50,
+              ),
+              Align(
+                  alignment: Alignment.bottomCenter,
+                  child: Column(children: [
+                    TextButton(
+                      child: Text(
+                        'Confirm',
+                        style: TextStyle(fontSize: 20.0, color: MyColors.white),
+                      ),
+                      style: TextButton.styleFrom(
+                        backgroundColor: MyColors.blue,
+                        fixedSize:
+                            Size.fromWidth(MediaQuery.of(context).size.width),
+                        shape: const RoundedRectangleBorder(
+                            borderRadius:
+                                BorderRadius.all(Radius.circular(10))),
+                      ),
+                      onPressed: () {
+                        if (_formKey.currentState!.validate()) {
+                          _formKey.currentState!.save();
+
+                          BlocProvider.of<UpdateEmailByCodeBloc>(context).add(
+                              UpdateEmailByCodeProfileEvent(
+                                  code: code.text,
+                                  emailToken: emailToken,
+                                  context: context));
+                        }
+                      },
+                    ),
+                  ]))
+            ],
+          ),
+        ));
   }
 
   _buildEmailbody() {
@@ -228,39 +234,33 @@ class _UpdateEmailWidgetState extends State<UpdateEmailWidget> {
                 ),
               ),
               SizedBox(
-                height: 5,
-              ),
-              SizedBox(
                 height: 50,
               ),
               Align(
                   alignment: Alignment.bottomCenter,
                   child: Column(children: [
-                    Container(
-                      width: width(context),
-                      decoration: BoxDecoration(
-                        color: MyColors.blue,
-                        borderRadius: BorderRadius.all(
-                          Radius.circular(20),
-                        ),
+                    TextButton(
+                      child: Text(
+                        'Confirm',
+                        style: TextStyle(fontSize: 20.0, color: MyColors.white),
                       ),
-                      //margin: EdgeInsets.all(25),
-                      child: FlatButton(
-                        child: Text(
-                          'Confirm',
-                          style: TextStyle(fontSize: 20.0),
-                        ),
-                        textColor: MyColors.white,
-                        onPressed: () {
-                          if (_formKey.currentState!.validate()) {
-                            _formKey.currentState!.save();
+                      style: TextButton.styleFrom(
+                        backgroundColor: MyColors.blue,
+                        fixedSize:
+                            Size.fromWidth(MediaQuery.of(context).size.width),
+                        shape: const RoundedRectangleBorder(
+                            borderRadius:
+                                BorderRadius.all(Radius.circular(10))),
+                      ),
+                      onPressed: () {
+                        if (_formKey.currentState!.validate()) {
+                          _formKey.currentState!.save();
 
-                            BlocProvider.of<UpdateEmailBloc>(context).add(
-                                UpdateProfileEmailEvent(
-                                    newEmail: neEmail.text, context: context));
-                          }
-                        },
-                      ),
+                          BlocProvider.of<UpdateEmailBloc>(context).add(
+                              UpdateProfileEmailEvent(
+                                  newEmail: neEmail.text, context: context));
+                        }
+                      },
                     ),
                   ]))
             ],
