@@ -1,25 +1,37 @@
+import 'package:clean_app_sample/core/material/colors.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
 class AlertMessage {
   alert(BuildContext context, String message) {
-    return AlertDialog(
-      title: Text(
-        'Message',
-        // style: TextStyle(fontSize: 12),
+    return ScaffoldMessenger.of(context).showSnackBar(
+      SnackBar(
+        backgroundColor: MyColors.blue,
+        content: Text(
+          message,
+          style: const TextStyle(color: Colors.white, fontFamily: 'Cairo'),
+        ),
+        shape: const RoundedRectangleBorder(
+            borderRadius: BorderRadius.only(
+                topRight: Radius.circular(20), topLeft: Radius.circular(20))),
       ),
-      content: Text(
-        message,
-        style: TextStyle(fontSize: 16),
+    );
+  }
+
+  loading(BuildContext context) {
+    return ScaffoldMessenger.of(context).showSnackBar(
+      SnackBar(
+        backgroundColor: MyColors.blue,
+        content: Row(
+          children: [CircularProgressIndicator(), Spacer()],
+        ),
+        shape: const RoundedRectangleBorder(
+          borderRadius: BorderRadius.only(
+            topRight: Radius.circular(20),
+            topLeft: Radius.circular(20),
+          ),
+        ),
       ),
-      actions: [
-        TextButton(
-          child: Text("OK"),
-          onPressed: () {
-            Navigator.pop(context);
-          },
-        )
-      ],
     );
   }
 }
